feat: храним баллы в формате decimal

+++ internal/dto/withdrawal.go
@@ -1,17 +1,17 @@
 package dto
 
 type WithdrawalRequest struct {
-	OrderID string `json:"order"`
-	Sum     int    `json:"sum"`
+	OrderID string  `json:"order"`
+	Sum     float64 `json:"sum"`
 }
 
 type BalanceResponse struct {
-	Current   int `json:"current"`
-	Withdrawn int `json:"withdrawn"`
+	Current   float64 `json:"current"`
+	Withdrawn float64 `json:"withdrawn"`
 }
 
 type WithdrawalResponse struct {
-	OrderID     string `json:"order"`
-	Sum         int    `json:"sum"`
-	ProcessedAt string `json:"processed_at"`
+	OrderID     string  `json:"order"`
+	Sum         float64 `json:"sum"`
+	ProcessedAt string  `json:"processed_at"`
 }
